feat[staging-provider]: Split Workspace and WorkspaceWithCredentials
