+++ libs/staging-provider/src/Staging/Workspace/WorkspaceStagingInterface.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace Keboola\StagingProvider\Staging\Workspace;
 
+use Keboola\StagingProvider\Staging\StagingInterface;
 use Keboola\StagingProvider\Workspace\WorkspaceInterface;
 
-interface WorkspaceStagingInterface extends WorkspaceInterface
+interface WorkspaceStagingInterface extends WorkspaceInterface, StagingInterface
 {
-    public function cleanup(): void;
+
 }
