+++ libs/staging-provider/src/Workspace/Credentials/CredentialsProvider.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Keboola\StagingProvider\Workspace\Credentials;
 
 use Keboola\StagingProvider\Workspace\Configuration\WorkspaceCredentials;
-use Keboola\StagingProvider\Workspace\Workspace;
+use Keboola\StagingProvider\Workspace\WorkspaceInterface;
 
 class CredentialsProvider implements WorkspaceCredentialsProviderInterface
 {
@@ -14,7 +14,7 @@ class CredentialsProvider implements WorkspaceCredentialsProviderInterface
     ) {
     }
 
-    public function provideCredentials(Workspace $workspace): array
+    public function provideCredentials(WorkspaceInterface $workspace): array
     {
         return $this->userProvidedCredentials->credentials;
     }
