+++ libs/staging-provider/src/Workspace/Credentials/ResetCredentialsProvider.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Keboola\StagingProvider\Workspace\Credentials;
 
 use Keboola\StagingProvider\Workspace\SnowflakeKeypairGenerator;
-use Keboola\StagingProvider\Workspace\Workspace;
+use Keboola\StagingProvider\Workspace\WorkspaceInterface;
 use Keboola\StorageApi\WorkspaceLoginType;
 use Keboola\StorageApi\Workspaces;
 
@@ -17,7 +17,7 @@ class ResetCredentialsProvider implements WorkspaceCredentialsProviderInterface
     ) {
     }
 
-    public function provideCredentials(Workspace $workspace): array
+    public function provideCredentials(WorkspaceInterface $workspace): array
     {
         if (in_array($workspace->getLoginType(), [
             WorkspaceLoginType::SNOWFLAKE_SERVICE_KEYPAIR,
