+++ libs/staging-provider/src/Workspace/Credentials/WorkspaceCredentialsProviderInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Keboola\StagingProvider\Workspace\Credentials;
 
-use Keboola\StagingProvider\Workspace\Workspace;
+use Keboola\StagingProvider\Workspace\WorkspaceInterface;
 
 interface WorkspaceCredentialsProviderInterface
 {
-    public function provideCredentials(Workspace $workspace): array;
+    public function provideCredentials(WorkspaceInterface $workspace): array;
 }
