+++ libs/staging-provider/src/Workspace/WorkspaceInterface.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Keboola\StagingProvider\Workspace;
 
-use Keboola\StagingProvider\Staging\StagingInterface;
+use Keboola\StorageApi\WorkspaceLoginType;
 
-interface WorkspaceInterface extends StagingInterface
+interface WorkspaceInterface
 {
     public function getWorkspaceId(): string;
 
@@ -14,20 +14,5 @@ interface WorkspaceInterface extends StagingInterface
 
     public function getBackendSize(): ?string;
 
-    /**
-     * @return array{
-     *      container?: string|null,
-     *      connectionString?: string|null,
-     *      host?: string|null,
-     *      warehouse?: string|null,
-     *      database?: string|null,
-     *      schema?: string|null,
-     *      user?: string|null,
-     *      password?: string|null,
-     *      privateKey?: string|null,
-     *      account?: string|null,
-     *      credentials?: array|null,
-     * }
-     */
-    public function getCredentials(): array;
+    public function getLoginType(): WorkspaceLoginType;
 }
